removed buildTxToSign function and did number to hex inside signAndBroadcast
